requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-base64==2.0
drf-extra-fields==3.4.1
psycopg2-binary==2.8.6
pytz==2022.6
sorl-thumbnail==12.9.0
sqlparse==0.3.1
tzdata==2022.7
